Add Player.py module
Refactored player logic by moving the Robot class to a new module, Player.py, in the actors folder.
